+++ apps/api/src/server.ts
@@ -31,7 +31,7 @@ import collaboratorsRouter from '@/routes/collaboratorsRouter.js';
 import { serverConfig } from './config/serverConfig.js';
 import logger from './logger.js';
 import authRouter from './routes/authRouter.js';
-import signatureRouter from './routes/signature-router.ts';
+import signatureRouter from './routes/signatureRouter.ts';
 import sessionMiddleware from './session/sessionMiddleware.js';
 
 const __filename = fileURLToPath(import.meta.url);
